Enable compiler warnings (#146)
* seperate ngstents into its own target, link that to ngstrefftz, and enable compiler warnings on ngstrefftz

* cleanup

* fix unused warnings

* fix unused warnings

* fix comparison of different integer types

* add missing include

* fix imports, fix unused variables

* fix unused parameter warnings

* simplify

* fix unused parameter warnings

* fix unused imports

* fix missing imports, fix unused imports, fix unused parameters

* fix unused parameters

* fix unused parameters

* fix unused parameters, fix unused imports, fix missing imports

* fix comparison warning

* add missing virtual destructor

* differentiate between clang and gcc, allow VLAs

* change library type for windows build

* make gcc happy

* try to fix windows builds

* do windows builds the old way

* add tent includes to windows build

* separate windows and non-windows-build

* update ngstents, switch back to Paul's repo

* revert to compiling all in one

* switch to fixed ngstents version

* enable more clang compiler warnings

* remove tautology

---------

+++ src/condensedg.cpp
@@ -49,7 +49,7 @@ namespace ngcomp
         ma->GetFacetElements (i, elnums);
         for (int elnr2 : elnums)
           {
-            if (elnr2 != elnr.Nr ())
+            if (elnr2 != (int)elnr.Nr ())
               els.Append (elnr2);
           }
       }
@@ -105,7 +105,7 @@ namespace ngcomp
                 Array<int> els2 = GetElNeighbours (ma, ElementId (elnr));
                 for (auto elnr2 : els2)
                   {
-                    if (elnr2 == ei.Nr () || els.Contains (elnr2))
+                    if (elnr2 == (int)ei.Nr () || els.Contains (elnr2))
                       continue;
                     els.Append (elnr2);
                     dnums.SetSize0 ();
